fix(backfire): fp32 multi-backend load; name Creative Mayhem UG as copyright holder
Force every diffusion backend to fp32 on load. Distilled models such as tiny-sd ship
fp16 components that mismatched the fp32 pipeline in cross-attention and crashed the
--models path; the single-backend default was unaffected.

Name Creative Mayhem UG as the copyright holder and commercial-license grantor in
LICENSE, LICENSING.md, and every source header, so the owner and the grantor are one
legal entity across the watermark IP. Set the licensing and legal contacts to
licensing@creativemayhem.com and legal@creativemayhem.com.

[skip ci]

+++ python/backfire/backfire.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: AGPL-3.0-or-later OR LicenseRef-Backfire-Commercial
+# Copyright (C) 2026 Creative Mayhem UG (haftungsbeschränkt)
 """Backfire: an imperceptible, keyed image watermark that AI provenance-stripping
 attacks *amplify* instead of remove.
 
@@ -98,10 +99,12 @@ def embed_cmd(a):
 
     def build_purifier(model_id):
         log(f"loading diffusion purifier proxy: {model_id} ...")
-        vae = AutoencoderKL.from_pretrained(model_id, subfolder="vae").to(dev).eval()
-        unet = UNet2DConditionModel.from_pretrained(model_id, subfolder="unet").to(dev).eval()
+        # force fp32 on every backend: some models (e.g. distilled ones) ship fp16
+        # components, which mismatch the fp32 pipeline in cross-attention.
+        vae = AutoencoderKL.from_pretrained(model_id, subfolder="vae").to(dev).float().eval()
+        unet = UNet2DConditionModel.from_pretrained(model_id, subfolder="unet").to(dev).float().eval()
         tok = CLIPTokenizer.from_pretrained(model_id, subfolder="tokenizer")
-        txt = CLIPTextModel.from_pretrained(model_id, subfolder="text_encoder").to(dev).eval()
+        txt = CLIPTextModel.from_pretrained(model_id, subfolder="text_encoder").to(dev).float().eval()
         sched = DDIMScheduler.from_pretrained(model_id, subfolder="scheduler")
         for m in (vae, unet, txt):
             for p in m.parameters(): p.requires_grad_(False)
